refactor: move non-specific renderer functions into dom_utils in new namespace
