+++ tests/attempt_ui/question_uis/format-floats.xhtml
