+++ tests/attempt_ui/question_uis/shuffle.xhtml
