+++ tests/question_test.php
@@ -21,11 +21,11 @@ use moodle_exception;
 use qbehaviour_questionpy;
 use qtype_questionpy\api\api;
 use qtype_questionpy\api\package_api;
+use qtype_questionpy\attempt_ui\question_ui_metadata_extractor;
 use qtype_questionpy\event\grading_response_failed;
 use qtype_questionpy\event\starting_attempt_failed;
 use qtype_questionpy\event\viewing_attempt_failed;
 use qtype_questionpy_question;
-use question_attempt;
 use question_bank;
 use question_engine;
 use question_state;
